resolve rebase probleme
merge

update automaticaly the device id of sensor that battery as been replaced.

algo

if a new device is detected with batery level 100% (new batery)
 we search if a same device type/subtype/Unit already exist with batery low
if one is found , we assume that it is the same device that we have replace batery and a new ID as been modified.
the old DeviceID is replaced by the new DeviceID
to allow this , the settings option  other/"Allow Device Replace On Battery Changed Detected" shall be ON